+++ src/server/game/Entities/Creature/CreatureGroups.cpp
@@ -113,7 +113,7 @@ void FormationMgr::LoadCreatureFormations()
 
         //Load group member data
         FormationInfo member;
-        member.LeaderSpawnId              = fields[0].GetUInt32();
+        member.LeaderSpawnId              = fields[0].GetUInt64();
         ObjectGuid::LowType memberSpawnId = fields[1].GetUInt64();
         member.FollowDist                 = 0.f;
         member.FollowAngle                = 0.f;
